+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 2025-10-31
 - #2002 **Resync breaking change**. This PR moves rollup configuration files to demo-rollup/configs directory. 
+- #2007 Changes celestia rollup metrics, please switch to new dashboard.
 
 # 2025-10-29
 - #1996 Set gas limit to 1B on ETH API access.